[FIX] HDF5 decoding of bytestrings (#332)
fix bytestring decoding

+++ examples/datasets/plot_data_beyeler2019.py
@@ -32,6 +32,7 @@ The data itself will be provided as a Pandas ``DataFrame``:
 # sphinx_gallery_thumbnail_number = 2
 
 from pulse2percept.datasets import fetch_beyeler2019
+
 data = fetch_beyeler2019()
 print(data)
 
@@ -106,8 +107,10 @@ argus = ArgusII(x=-1331, y=-850, rot=-0.495, eye='RE')
 # For now, we add the necessary columns ourselves.)
 import pandas as pd
 data = fetch_beyeler2019(subjects='S2')
-data['img_x_dva'] = pd.Series([(-30, 30)] * len(data), index=data.index)
-data['img_y_dva'] = pd.Series([(-22.5, 22.5)] * len(data), index=data.index)
+data['img_x_dva'] = pd.Series([(-30, 30)] * len(data), index=data.index,
+                              dtype=float)
+data['img_y_dva'] = pd.Series([(-22.5, 22.5)] * len(data), index=data.index,
+                              dtype=float)
 
 ###############################################################################
 # Passing both ``data`` and ``argus`` to
@@ -161,4 +164,4 @@ plt.xlabel('phosphene elongation')
 
 ###############################################################################
 # Phosphenes are not pixels!
-# And we have just reproduced Fig. 3C of [Beyeler2019]_.
+# And with that we have just reproduced Fig. 3C of [Beyeler2019]_.
